Create BST.cpp
A complete binary tree is a binary tree whose, all levels except the last one are completely filled, the last level may or may not be completely filled and Nodes in the last level are as left as possible. we have given the root of a complete binary tree. in this program we are finding the count of nodes.